shared perf monitor
